+++ src/traits.rs
@@ -1,31 +1,17 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
-use cedar_policy::{Entity, EntityUid, RestrictedExpression};
+use cedar_policy::{EntityUid, RestrictedExpression};
 
 use crate::error::PolicyError;
 
 /// Anything that can become a Cedar‐typed atom, e.g. `User::"alice"`,
 /// `Action::"foo"`, `Group::"devs"`.
 ///
-/// Types implementing this trait can produce their Cedar identity (`cedar_id`),
-/// attributes, and parent relationships for evaluation.
-pub trait CedarAtom {
+/// Types implementing this trait can produce their Cedar identity (`cedar_id`)
+/// and attributes for internal request preparation.
+pub(crate) trait CedarAtom {
     /// The Cedar typename (“User”, “Action”, “Group”, etc)
     fn cedar_type() -> &'static str;
-
-    /// Build a Cedar parent list for this atom, default is no parents.
-    fn cedar_parents(&self) -> HashSet<EntityUid> {
-        // Default: no parent type
-        HashSet::new()
-    }
-
-    /// Build an entity for this Cedar atom, e.g. `User::"alice"` or `Host::"flappa.example.com"`.
-    ///
-    fn cedar_entity(&self) -> Result<Entity, PolicyError> {
-        let entity_uid = self.cedar_entity_uid()?;
-        let attrs = self.cedar_attr()?;
-        Ok(Entity::new(entity_uid, attrs, self.cedar_parents())?)
-    }
 
     /// Build the attributes for this Cedar atom.
     fn cedar_attr(&self) -> Result<HashMap<String, RestrictedExpression>, PolicyError> {